Add a script to fix ripemd160
Hashlib uses OpenSSL for ripemd160 and apparently OpenSSL disabled some older
crypto algos around version 3.0 in November 2021. All the functions are still
there but require manual enabling. But we use ripemd160 for tests.

For the system tests on GitHub runners to change the OpenSSL configuration
files, a script and tests for it have been added.
Of course, the user can also run this script on servers or laptops, but it's
main purpose is to prepare the environment on ubuntu-latest runners.

Signed-off-by: Oleg Kulachenko <oleg@nspcc.ru>