testsuites: Create temporary directories
Create temporary directories for keeping files and objects.
In the process of running a full test suite, a lot of test files and objects
are generated. About 50 Gb, that is not enough space available on the
ubuntu-latest runners.

Some tests are written to reuse test files and objects created by previous
tests in the test module. For this reason, this commit introduces functionality
to create separate directories for files and objects and to remove
the directories when the test module is finished.

This commit is part of the bug fixes
https://github.com/nspcc-dev/neofs-testcases/issues/571

Signed-off-by: Oleg Kulachenko <oleg@nspcc.ru>