tools: Add a script for deploying Allure reports to NeoFS #574
https://github.com/nspcc-dev/neofs-node/issues/2324

Added a script for deploy Allure tests result as static
web page to NeoFS.
All zip archives with logs are stored as different objects.
This script is planned to be used in GitHub actions.

This new functionality will help to save any test report as an
object in the NeoFS network, we will become independent of
the limitations of the GitHub branches and pages.

It will also be an "advertisement" for our object storage.

Signed-off-by: Oleg Kulachenko <oleg@nspcc.ru>