tests: decrease amount of created objects for load test
This is not a real load test, so creating a high number of objects
doesn't make sense. 250 objects can sometimes leads to a timeout inside ssh
command. So the amount of objects was decreased to 100. Decreasing
further - to 2 objects seems to be rather dangerous, cause some
bugs can be missed.

Signed-off-by: Evgeniy Zayats <zayatsevgeniy@nspcc.io>