tests: handle OOM during allure logs attachment
As a result of test files layout update, we tried to pack too many files to
allure report in case of test failures. During allure logs attachment a
copy of all existent files is made by allure plugin.
Taking into account parallel tests runs in CI  - it caused OOM.
In this commit amount of files for attachment is reduced and a removal
of storage related files is added.

Signed-off-by: Evgeniy Zayats <zayatsevgeniy@nspcc.io>