tests: adjust numerical tolerance for payments tests
Expected value for the single SN debit operation is single_gain_per_epoch, but in
reality it never matches exactly, there is a small difference. The minimum allowed difference was
0.5, but sometimes it is a little bit higher (e.g. 0.555), so to tolerate such errors we increase
this allowed difference to 1, which still seems quite reasonable. The alternative is to remove this
assertion at all, but then we can miss some unexpected significant charge or debit operations.

Signed-off-by: Evgeniy Zayats <zayatsevgeniy@nspcc.io>