tests: remove post-drop access check for ec recovery test
The check is to verify that the drop actually works.
But the test actually verifies the ec recovery behavior.
We have other tests to verify that the drop functionality works.
Otherwise, we would need to have a separate env with a custom policer
and we will need to restart all SNs after an object drop to update the
policer settings.

closes #1236

Signed-off-by: Evgeniy Zayats <zayatsevgeniy@nspcc.io>